Make SyncSendPointer unsafe and add HostObjectPointer
Constructing a `SyncSendPointer` implies that the object pointed to by
it as actually `Sync` and `Send`. Updated docs and marked unsafe to
reflect that.

In most places we were using `SyncSendPointer`, the pointer is *not*
actually fully thread-safe, and we instead rely on the Host lock being
held at access time. Created `HostObjectPointer` for such pointers.