Run actions on pull_request_target instead of pull_request
This causes the actions to run on the head of the pull request instead
of on the merge result. This is generally what we want, particularly
since we require PRs to be up to date with the target branch before
merging anyway.

This also causes code to run in the context of the pull request branch.
If our actions used secrets this would cause them not to get them, or if
they tried to manipulate the repo they wouldn't be able to, etc. This
should be ok since these don't use secrets, and we probably don't *want*
a PR to be able to inject malicious code into a PR that can use our
permissions. (We already generally drop permissions, but even better if
the action never has them in the first place.)

Fixes https://github.com/shadow/shadow/issues/2166