Use longer timeout for coverage testing (#2076)
In particular `select-shadow-ptrace` has been timing out in coverage builds occasionally.

We could alternatively just bump the timeout of that test, but it's a little unfortunate to set the timeouts based on how long they take in coverage builds, which is much longer than otherwise.

This change won't affect tests with overridden timeout properties. We could address that by making a Coverage CMake build type and explicitly setting the timeout for all tests (with a multiplier in coverage builds), but doesn't seem worth it right now.