shim_tls: lazily allocate TLS for each thread
This removes the hard-coded limit on the number of threads, and avoids
allocating memory for threads that are never actually created.

Unfortunately the TLS for exited threads still leaks (as it did before).
We could do something like catch in shim_native_syscallv that we're
about to call SYS_exit (which is the only way for a thread to exit
without exiting the whole process), but we can't deallocate the thread's
TLS from there since it's running on stack allocated in that TLS.

We could perhaps add the TLS to a 'free list' to be freed or recycled by
another thread, but leaving that for now.