Free thread signal stacks
We can't free them from a thread that's exiting, since it is likely
running on the stack at the time.

Instead we save a pointer to a global for another thread to free.  The
next time another thread exits, it'll free the pending stack before
saving its own pointer to be freed.