Warn if trying to add syscall counts when there's no worker
This doesn't seem to happen in practice, so panics when a debug build.
Previously this used the global manager's syscall counts when there was no
worker, but we don't need this complexity, and if we actually do try to add
syscall counts when there's no worker, then it's probably an error anyway.