Pre-allocate space in shared memory files (#2267)
Replace usage of `ftruncate` with `posix_fallocate`. `ftruncate` sets
the size of a file without actually allocating the backing storage,
which is allocated on-demand. While this allows the true allocation to
be "lazy", and thereby allows us to do large allocations without wasting
resources if the whole allocation isn't used, it can cause unpredictable
and confusing failure modes.

With `posix_fallocate`, the space is eagerly allocated. If the call
succeeds, we are guaranteed that the storage is actually available.
When we run out of storage, the call fails and we can fail with a clear
error message.

We were previously allocating quite-large chunks of memory at a time;
about 132 MB. While this probably wouldn't actually cause problems for a
single shadow simulation (which typically only needs one such
allocation), it ends up using a lot of space when running the shadow
tests in parallel. This changes the default allocation to 1 MB, instead.

Fixes https://github.com/shadow/shadow/issues/2266