run_tests_incremental: minor cleanup (#2320)
* run_tests_incremental: quote FLAGS lists

I thought these needed to *not* be quoted to be interpreted as separate
flags, but that appears not to be the case. I guess the rules are
different for list-expansions.

Afaict this doesn't actually change the behavior in any way, but in
general *not* quoting variables in bash is a bit of a red flag, so
better to quote if we can.

* run_tests_incremental: turn on bash command echoing

We run a somewhat complex bash block, with conditional logic, as a
single step. It's probably useful to see what individual commands are
actually running.