SysCallReturn: make a proper tagged union
Adds interfaces to SysCallReturn to ensure that they are constructed
and accessed consistently with their type.

This also adds structs to represent each individual potential state of
the SysCallReturn, which will be helpful for a follow-on refactor to
move more of the syscallhandling code from ManagedThread to Thread.

While updating the syscall handlers I perhaps-ill-advisably fixed some
handlers to not check the result of the actual memory access operation
instead of just checking for NULL, which led to needing to fix some
other minor bugs:

* In test_connect, when testing a NULL pointer, I ensured that other
arguments are consistent, to remove ambiguity over whether EINVAL or
EFAULT should be expected.
* In the memory manager, if we try and fail to flush a mutable
reference, still mark the buffer as non-dirty, so that we don't panic
when the buffer is dropped.