Pull in objgraph via Cargo instead of a top-level cmake project
I haven't been able to get cmake to respect dependencies on the separate
objgraph project, which in clean builds results in the shadow-c build
failing, since it tries to link it without it being built first.

Also replace the Host lock with an objgraph Root, mostly to validate
that we're able to link it correctly.