Changed host shuffle method and moved it to the manager
We'll want to change the shuffle method to a rust method at some point, and by
changing only the shuffle method now (the shuffle is changed from C to rust),
we can isolate simulation performance changes due to the different host order
and not conflate it with other changes.

I also think shuffling in the manager makes more sense anyways.