Moved some of `Controller`'s global state to the worker module
The controller is currently globally accessible through the global worker pool,
but since the controller exists on the stack, this is not allowed in rust. This
breaks out some of the global state into a new global object. Accesses to this
global state is made explicit through a `worker::WORKER_SHARED` static
variable.