Add shadow-build-common
This will be used by the build.rs scripts to build C and C++ code, using
consistent flags and settings.

Some flags specified in the cmake files are explicitly specified here,
but we currently also accept some flags via env variables so that we
don't need to implement all of the conditional logic yet.