Run rust unit tests (including doc tests) (#2421)
Accidentally dropped running any rust unit tests in
https://github.com/shadow/shadow/pull/2404

We were also previously not running doc tests, and I think not running
tests for any crates for which we don't build a `staticlib` for C code
to link to. (I'm not sure whether there were any such tests before, but
there increasingly will be)

For simplicity and maintainability I just invoke `cargo test` now,
instead of trying to directly run the test executables. While the cargo
behavior of always rebuilding (if needed) may be slightly surprising,
this approach should be more robust and maintainable.