bindgen and cbindgen: put generated files in build dir
This eliminates spurious rebuilds between consecutive calls to `cargo
test` (./setup test rust-unit-tests). Since the generated bindings were
in the source directory, it always detected that something in the source
directory had changed and so needed to rerun.

There's still a (spurious?) rebuild on the first
call to `cargo test` after the initial build, but I'm not sure why.