setup: Always run bear if present
Use `bear`'s `--append` flag to update an existing
`compile_commands.json`, so that we can run it on incremental builds as
well without erasing entries for files that weren't rebuilt.

With this option, it generally makes sense  to always run `bear` when
building, so this also removes the `--compile-commands` flag from
`setup`, and decides whether to use `bear` based on whether it appears
to be present on the system.