Host: move Rng from C to Rust
This allows us to remove the previously unsound api Host::random, which
returned a mutable reference from an immutable reference without using
interior mutability.

Rather than trying to expose the Rng directly to C, with the needed
function object, we implement the rng APIs that the C code needs
directly on the Host.

After this change we no longer need the Random C APIs, so those are
removed.