Small change to packet reliability calculation
Since 'chance' has the range [0,1) and reliability has the range [0,1], we
should be comparing 'chance' to 'reliability' with a '>=' instead of a '>'.
Otherwise it's possible for a packet to travel along an edge with 0%
reliability.

Edge case that didn't work before but now does:

- chance=0, reliability=0 -> 0 >= 0 -> true -> drop

Other edge case that did work before and continues to work:

- chance=*, reliability=1 -> * >= 1 -> false -> don't drop