shmem allocator: deallocate allocated blocks on Drop
The previous approach of requiring explicit deallocation effectively
forces structs containing an allocated block that it deallocates on Drop
to wrap in an Option, making that type a little more awkward to work
with.

The ability to *not* deallocate an allocated block on Drop was only
needed for blocks that were deserialized via the Allocator instead of
the Serializer. I'm not aware of a use-case for that functionality, so I
removed it.

Also:

* Remove the `unsafe` marker from `Allocator::alloc`. Since the blocks
  allocated this way are now always valid until the block is dropped,
  this doesn't need to be `unsafe`. (In contrast, blocks obtained via
  the serializer *are* unsafe, since it's the user's responsibility to
  ensure they're not used after the original block is dropped).

* Call T's Drop implementation when dropping a block, if applicable.

* Add a test validating that mutations via aliased blocks are observed
  by the other blocks.