Rewrite SelfContainedMutex to fix deadlock and test under loom
Fixes https://github.com/shadow/shadow/issues/2557

This collapses the lock state and the number of sleepers into a single
atomic, which we also use as the futex word. This makes the concurrency
correctness much easier to reason about, at the cost of having to do
some bit fiddling, and under contention sometimes having to do some
extra atomic operations to synchronize the sleeper-count.