SelfContainedMutex: move from shadow-shim-helper-rs to shadow_shmem
shadow_shmem is a better home for "utility" things to work with shared
memory that don't depend on the rest of Shadow.

This also lets us avoid having to compile shadow-shim-helper-rs when
testing SelfContainedMutex in the loom configuration, which saves some
work and lets us avoid some loom-specific workarounds in unrelated code.

Progress on https://github.com/shadow/shadow/issues/2566