run_tor workflow: build shadow once
Given infinite gitlab runners this doesn't improve latency, since the
end to end time is still (time to build shadow) + (time to build and
test tor), but it uses fewer github resources overall.

Since we seem to be bumping into limits of number of parallel jobs
permitted, this should help.