Build shadow once for extra tests
Previously we were building shadow for each version of tor that we test,
as well as to test examples and do bandwidth tests with tgen. These all
use the same build configuration and container, so we can save some GH
resources by putting these all in one workflow and building shadow once.

This also updates shadow configs and docs to use PATH-based search to
locate tgen and tor. It's not strictly necessary for this change, but
helps to decouple the "build shadow" job a bit from the other jobs that
use its output, and is something we're moving towards in general to
support alternative install locations.