Extra tests CI: improve Rust caching
* Cache more of ~/.cargo

Unclear whether this is needed, but matches other examples found online,
and doesn't seem to increase (compressed) cache size significantly.

* Add month to cache restore-keys

Using the month in the cache `restore-keys` in addition to `key`
ensures that we *reset* the cache every month, instead of continuing to
grow the previous month's cache.

* Include hash of the lockfile in cache keys but not restore-keys

This ensures we update the cache when our dependencies change.

* In extra_tests: Use sccache to cache built rust dependencies

Largely cargo-culted from
https://www.infinyon.com/blog/2021/04/github-actions-best-practices/#optimizing-rusts-build-speed-with-sccache

Only ends up using ~200 MB of cache, and speeds up the build quite a
bit.