Support `close()` for rust TCP wrapper
The `LegacyCallbackQueue` prevents circular borrows when running event
listeners. For the `close()` syscall, this prevents the panic where the syscall
handler mutably borrows the `TcpSocket`, the inner C `TCP` object changes to
the `CLOSED` state, an epoll event listener callback is run, and the callback
takes a shared borrow of the `TcpSocket`.

A temporary `LegacyCallbackQueue` is stored as a thread-local so that the
`LegacyDescriptor` can access it without us having to update large amounts of C
code to pass it through to everywhere we need it. Since we're
deprecating/removing the C code, it's not worth the effort / possible breakage
to update it.