Process: move DescriptorTable to Rust, with int mutability
To avoid recursive borrow problems, we'll for now need to *not* have a
mutable borrow of Process for the whole duration of syscalls, since
legacy C code needs to try to borrow again.

For that to work, we need to make Process's mut methods be non-mut,
which means we need interior mutability.

To implement interior mutability we need to move corresponding objects
into the Rust Process.