Remove usage of process_getWritablePtrRef from outside of Process
Using it directly this way skips the runtime checking that Process does
to ensure there are no incompatible borrows from the MemoryManager.

I think the primary reason it was being used here was so that errors
writing to memory could be handled instead of panicking. Now that
process_flushPtrs supports this, we can use that instead.