Process: move cached memory references from C to Rust
The C code was attempting to detect incompatible borrows, but had some
gaps. It detected incompatible borrows by the legacy "caching" memory
access APIs with each-other, but not with the newer APIs.

During this migration I cached the Ref or RefMut of the MemoryManager
used to create the memory reference as well as the memory reference
itself. This requires a self-referential data structure and is a bit
fragile, but should prevent incompatible borrows more reliably than the
old code.

As a result, I had to add some extra flushes to ensure there weren't
outstanding borrows from a syscall handler when running the
strace-logging code, which also tries to access memory.