Process: wrap threads in `RootedRc<RootedRefCell<_>>`
The `RootedRc` lets us hold a reference to a thread without holding a
reference to the thread list.

`ThreadRef` currently has mutable methods, so we need to be able to get
a mutable reference, so we use `RootedRefCell`. We could end up dropping
this if we change `ThreadRef` to use internal mutability everywhere as
we do with Process and Host.