Refill token buckets less often on slow connections
The token bucket now returns the time that the next packet conforms,
ie how long until we will have enough tokens to forward that packet.
Returning this instead of how long until the next refill reduces the
overall number of forwarding events we need to schedule, and is
especially helpful on slow connections.