Handling SHD_SHIM_EVENT_PROCESS_DEATH: mark whole process as dead
We inject a SHD_SHIM_EVENT_PROCESS_DEATH event into each individual
managed thread when we detect the process has died.

In https://github.com/shadow/shadow/issues/2674, this ended up causing
the memory write inside Process::reap_thread to fail. Since the process
has died, we don't need to do that write at all; marking the process as
dead correctly skips this step.

I tried to write a regression test but was unable to reproduce the
issue. As in the mongodb case, I set it up so that a child thread
called `abort` while the thread group leader was blocked. In both cases,
we end up trying to use the thread group leader to do the write, but it
is in a zombie state since we already natively raised SIGABRT, which
kills the whole process. In my test that write *works*, even though the
thread group leader is in a zombie state.

I did verify that this change fixes the problem in the mongodb example,
though.