Push local packets directly to netif, stop using events (#2711)
Local packets were previously "sent" to the netif by scheduling an
event. This commit changes this behavior to instead just directly
push local packets to the destination netif. This results in fewer
events (it saves 1 event for every local packet) and should thus be
more performant. It does change the order of operations a bit, so
we expect non-deterministic results across this commit.