Set shim's stdout/stderr to the shim log file
Shadow now uses an fd passed through fork/exec for the shim log rather than
passing a filename using `SHADOW_LOG_FILE`, and also uses this fd for stdout
and stderr.

If the shim writes to stdout/stderr (or the plugin disables interposition and
writes to stdout/stderr), it should be written to the shim log rather than to
shadow's stdout/stderr (which may be the terminal or some other file).