Thread::new: return RootedRc<RootedRefCell<Self>> instead of Self
Since in the planned design we don't intend to use mutable references to
Thread, I thought we could omit the RootedRefCell wrapper here. We
*could*, but without it we need to make Thread itself Sync, which means
that for interior mutability we end up needing to use e.g. RootedCell
instead of just Cell, and RootedRefCell instead of just RefCell. That in
turn requires callers of most of Thread's methods to provide a Root.

I think it'll be simpler (and maybe more performant) to just wrap the
Thread itself in RootedRefCell, in which case Thread only needs to be
Send, but doesn't need to be Sync.