Remove deprecated spin config options
Notably this removes the user-facing experimental options
`preload_spin_max` and `use_explicit_block_message`, and hard-wires the
current defaults of not spinning and not using an explicit block message
(which would be redundant with spinning disabled).

We've stopped using these options, and I haven't bothered reimplementing
them in the Rust version of `IPCData`.