move writer_closed to a separate atomic
This mostly works, but:

* there's a loom failure I can't figure out in one test (which
shouldn't actually matter in shadow)
* we can't safely drop the contents of the channel (if any)
after closing it. This might be fixable, but generally there's a
potential race if we try to drop the contents when closing, since it
could be getting read in parallel.

It appears to hurt performance without pinning, and be a no-op with
pinning.

```
$ cargo bench  --manifest-path=src/Cargo.toml  -p vasi-sync
ping pong               time:   [4.8376 µs 4.8934 µs 4.9665 µs]
                        change: [+10.053% +13.523% +17.508%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 11 outliers among 100 measurements (11.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild
  8 (8.00%) high severe

ping pong pinned        time:   [3.2015 µs 3.2051 µs 3.2094 µs]
                        change: [+0.5280% +0.6995% +0.8895%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```