Moved the `SysCallCondition`s active file assignment
I think it makes more sense to set the `SysCallCondition`s active file at the
top-level of each syscall handler rather than putting it in helpers. This means
that we don't need to pass the `OpenFile` all the way through, and means that
any helpers or the file methods don't need to worry about setting this
themselves when they want to return a blocking `SyscallError`.