ManagedThread: save notification handles when registering callbacks
We were failing to save the ChildPidWatcher notification handles, which
in turn meant we weren't unregistering threads with the ChildPidWatcher
after they exited, thus creating a resource leak. The relevant data was
still getting freed when the whole Process exited, so it would have only
affected Processes that had many threads start and end over their
lifetime.

The callback registered for each thread is only about 16 bytes - a
function pointer and a data pointer.