Use rustfmt to re-group imports
In previous discussion we agreed in theory on rustfmt's
`StdExternalCrate` grouping policy. Unfortunately we currently can't
use this in our default config, since it only works on nightly.

This is a one-time run of rustfmt using that policy with the nightly
compiler.

It cleans up the ordering both from the previous commit (which was
semi-automated using sed) and everywhere else.