ChildPidWatcher: fix a race condition when unregistering a pid (#2814)
We were getting rare failures in CI unit tests, and consistent failures
in large benchmarks. I was able to reproduce with:

```
for i in `seq 1000`; do
  echo trial $i
  if ! RUST_BACKTRACE=full cargo test --manifest-path=src/Cargo.toml -p shadow-rs childpid_watcher ; then
    break
  fi
done
```

I think what was happening is that we could get an epoll event for a
dying process in the worker thread just as we were in the middle of
unregistering it from another thread.

The simplest fix is to just let the worker thread handle unregistration
so that it's serialized with other operations in that thread.