CI: run containerized tests on latest ubuntu base image
Some of these were intentionally running on the oldest available base
image to try to test on an older kernel. Since it turns out GitHub's
Ubuntu base images use the same kernel version, we might as well run on
the newest one.

Similarly, since all of these tests run in containers, it's probably not
worth pinning the base image to a specific version of Ubuntu.