Move and better-document waiting for thread exit
This moves waiting for thread-exit from Process to ManagedThread, where
it more logically belongs. It also adds a check in ManagedThread's Drop
implementation to validate that we never drop it while the native thread
is still running, and adds documentation about why this is important -
to prevent memory corruption due to shared memory regions being
deallocated and reused.

This also fixes a bug - Before we *weren't* waiting for the thread to
exit if the `clear_child_tid` attribute wasn't set. I belive this was a
race condition, but in practice I think most higher level thread APIs
(e.g. libc's `pthread` and Rust's `std::thread`) use this attribute, so
wouldn't have wouldn't have been affected.