Move methods from Process to individual states
In most cases it's preferable for callers to get the process state and
then call methods directly on it, instead of proxying through the outer
Process. This gives control to the caller for how to handle unexpected
states, and avoids needing to use NestedRef and NestedRefMut, which
require heap allocations.