FinalProcessState: use nicer serialized format
Explicit yaml types can be difficult to work with. e.g. it's unclear how
to get pyyaml to emit `!running` (and not `"running"` or `!running {}`.
Even if we solved it for pyyaml, users are likely to run into problems
with other tools.

With a few tweaks we get a more intuitive yaml representation, at the
expense of a slightly more complex internal representation.