MemoryCopier: Use live thread ID for writes
In particular this fixes a crash when reaping a thread and trying to
write to the `child_tid` address, since the current thread is no longer
live and can't be used with the `process_vm_writev` syscall.

This problem was mostly latent, since when the memory mapper is enabled,
we usually do the write through shared memory instead.