Trying to migrate over; running into issues
Current blocker: since the types are wrappers around the original C
types, cbindgen wants to define them back in terms of those types. But,
including those kernel type definitions tends to be problematic and lead
to conflicts.

We could export them as opaque types, but then they wouldn't be sized,
so couldn't be copied etc.

We might be able to export as a sized opaque type, but we need to know
the size and alignment of the rust type, which isn't available when the
build script is running. Maybe we could get around that by having a
separate crate that generates the C headers after the original crate is
built.

Probably the most straight-forward fix is to go back to carefully
creating "parallel" definitions in the Rust code rather than directly
wrapping the C types. At least with the C bindings we can validate that
the size and alignment is correct (and maybe field offsets?).