syscall_handler: log only at warning for unimplemented syscalls
"error" is too noisy for unimplemented syscalls. They do not themselves
cause shadow to exit with an error, so they should only be a warning.

In many cases the managed program is able to fall back to some
alternative syscall, and if not then typically it will fail in an
obvious way; e.g. exit with a non-zero exit code, which *will* cause
shadow to report an *error* if that wasn't expected.