In clone: restore RDI and use ret to jump
Instead of "sacrificing" RDI to store our jmp address,
we can push it onto the stack and then use `ret` to do the
jump.

This could run afoul of some stack protection mechanisms (shadow
stacks?), but such a failure is likely to be easier to track down
than one due to not restoring RDI when the caller expects it to be
restored.