Allow-list seccomp native syscalls by code section
There's no easy way to get an authoritative "end" address of an
arbitrary individual function, so we were just guessing a loose upper
bound, and hoping that managed code that tried to make direct syscalls
didn't unluckily end up in the allow-listed address range.

We *can* reliably get the end address of a linker section, and place
functions we want to allow-list into that section.

Since we're not using a linker script, we're depending on the linker
to create the `__start_*` and `__end_*` symbols. This is a little
fragile, but should be obvious if it breaks, and resolvable by
adding/modifying a linker script to explicitly add such symbols
ourselves.