clone handler: handle individual flags more explicitly
This is mostly a refactor in preparation for supporting fork.
https://github.com/shadow/shadow/issues/1987

There are also a couple minor behavioral changes, that shouldn't affect
any of our supported thread libraries:

* Some flags that we don't actually care about are no longer required
(e.g. CLONE_SYSVEM).
* CLONE_SETTLS *is* now required. Previously we would have tried to go
ahead and execute the native clone without it, but the shim would
misbehave since it depends on thread local storage.