Pass thread and process serialized shmem in Start event
This detangles some initialization dependencies, so that we can create a
fully create a ManagedThread, then a Thread to contain it, then a
Process to contain it. It simplifies the existing code, and the
implementation of fork.

EDIT while rebasing: This commit incorrectly reinitializes the Process
shmem block (in addition to the Thread shmem block) when starting a new
thread.