Add the `LegacyTcp` wrapper to the interface qdisc
Previously the C `TCP` object would be added to the qdisc, but we want the rust
`LegacyTcp` wrapper to be added instead. This requires a circular reference
between the `TCP` and `LegacyTcp` objects since the `TCP` object adds itself to
the qdisc, but the `TCP` object has been made to only hold a weak reference to
the `LegacyTcp`. Some additional changes needed to be made to extend the
lifetime of the `LegacyTcp`, otherwise upgrading the weak reference won't work.