Let clone syscallhandler orchestrate construction
If we were to extend the current approach, we would've needed to
delegate all clone calls to Process, which delegates to Thread, which
delegates to ManagedThread, and handling of CloneFlags for emulated
behavior would have been scattered.

Instead we let the syscallhandler create the ManagedThread directly,
handle wrapping it in a Thread and Process, and continue handling the
logic for all CloneFlags.