SigInfo: be a bit more rigorous about safety
In particular implementing `bytemuck::TransparentWrapper` was a bit
questionable. safe Rust code might have been able to use the resulting methods
to create unsound behavior by initializing unexpected alternative fields
of the bindgen unions. (I think in practice the alternative fields are
padding, and are >= the size of the alternative field, s.t. initializing
it would still initialize all the bytes of the other fields... but this
seems fragile and could change out from under us).

Instead we manually impement `unsafe` versions of the methods we want
out of `bytemuck`, noting the invariants that must hold for it to
actually be safe (basically that the raw type wasn't initialized in a
bizarre way).

Some of the methods that access inner unions are also update to use
`addr_of` and `addr_of_mut` to avoid creating actual references that
might not be valid.