test_clone: Set empty thread-local-storage (TLS) descriptor
Previously, we didn't use `CLONE_SETTLS` at all. In this case the child
thread implicitly inherits the parent's thread-local-storage.

With this change, we explicitly *clear* thread-local-storage.

Natively, this will result in an easier-to-debug segfault if the child
thread tries to access native thread local storage, instead of accessing
its parent's thread local storage.

In Shadow's shim_tls, we detect this case and use a fallback for the
shim's own thread local storage.