vasi-sync: drop rkyv support
This was from an earlier attempt to explore using `rkyv` as a safe
abstraction to accessing shared memory, but it didn't really work out.

`rkyv` does support `no_std`, but it'd take a little work on our side
switch to that version of it. It doesn't make sense to bother since
we're not using this code anymore.

Progress on https://github.com/shadow/shadow/issues/2919