packet.c: fix packet logging in unit tests
This code was directly checking the `rustlogger` logging backend,
instead of going through the `logger` front-end. In unit tests this is
incorrect, since `rustlogger` is not installed, and `logger` instead has
the default `StderrLogger` backend.

Where the previous check against `rustlogger` returned `false` since no
Rust logging backend is installed in unit tests, going through the
frontend correctly returns `true`.

This casues us to log the packet status in unit tests, which is
a code path we weren't hitting before, and panicked since the `PMOCK`
case was unimplemented. This adds (~noop) handling for the `PMOCK` case
to fix the crash.