SelfContainedChannel: use guaranteed layouts
Since we put these in shared memory, we need `repr`s that guarantee a
consistent layout.

I made this change when debugging a memory error; it turned out to be
something else, but this change may help avoid future nasty surprises.

Ideally we'd enforce that types we use with the shared memory allocator
recursively have stable representations (repr(c) etc) by adding that to
the derive macro of one of our existing traits
(VirtualAddressSpaceIndependent) or a new one. I don't see a way to
actually check the repr in Derive macros, though...