shim TLS: use a proxy type for C allocation
We were previously restricted in what we could put inside
`ShimThreadLocalStorage`, since cbindgen needed to be able to generate
a C definition for it.

We actually only care about cbindgen creating a definition with the same
alignment and size though, so use a proxy type with the same alignment
and size.