Enable minimal optimizations in dev (debug) rust builds
Rust with no optimizations has a tendency to blow up stack usage, since
there is no "placement new" - even when large items are allocated in the
heap or static storage, initialization requires creating a temporary
version on the stack.

This can be problematic in the shadow shim in particular, where we
sometimes need to initialize relatively large objects, and sometimes
need to work in relatively small stack space.