Rework Rust version pinning
Replaces `rust-toolchain.toml` with `rust-toolchain-stable.toml` and
`rust-toolchain-nightly.toml`. This consolidates the pinning of the
nightly version (used in some CI tests) to one place.

For users/developers, this means the version won't be automatically
pinned anymore; they can choose to pin to the CI stable or nightly
version by creating a symbolic link to one of these two files.

Alternatively we could check in a symbolic link to pin to the stable
version by default, but then if a user wants to change the link to point
to the nightly version, it'd be easy to accidentally check in the new
link as well.