Change option log-errors-to-stderr to log-errors-to-tty
In particular, it's generally undesirable to log errors to both stdout
and stderr if both are actually the same destination. The new behavior
should always avoid that since it'll only ever log to stderr when it is
a tty but stdout is not; i.e. they can't be the same destination.

A possible downside of this approach is that it removes the ability to
log errors to stderr when stderr is a different destination than stdout
but not a tty.

This is probably a worthwhile tradeoff; particularly until any real
use-cases for wanting to log a non-tty stderr come up.