Support clone without `CLONE_FILES`
When `CLONE_FILES` is *not* set, the child gets a copy of the parent's
descriptor table instead of a reference to it. This is unusual but
permissable to do even when the `clone` syscall is not creating a new
process. i.e. threads within the same process can have different
descriptor tables.

To support this, the descriptor table must be moved from `Process` to
`Thread`. To continue to support the case where `CLONE_FILES` *is* set,
it must also be reference-counted. Accordingly, we wrap in a
`RootedRc<RootedRefCell<_>>`.