Implement Netlink sendmsg syscall
This commit implements the senmsg syscall.

Only the following parameters are supported:
- The destination pid must be kernel (pid=0), since the only entity that
  will reply the message is the kernel.
- There are no destination groups. Since the only destination is the
  kernel, multicasting does not make sense.
- The only supported flags are MSG_DONTWAIT, MSG_NOSIGNAL, and
  MSG_TRUNC.

Rate limiting is done by limiting the number of in-flight bytes by
NETLINK_SOCKET_DEFAULT_BUFFER_SIZE.

When the syscall is succesfully, the socket pushes the message into the
buffer and emits an event by using the StateEventSource instance.