Make process-spawn methods fallible
We're going to end up spawning a new ManagedProcess in the
implementation of execve. This should be fallible so that e.g. if a
managed process tries to exec a path that doesn't exist, we can just
return an error instead of crashing shadow itself.

Progress on https://github.com/shadow/shadow/issues/2988