Implement Netlink recvmsg syscall
This commit implements the recvmsg syscall.

The syscall handlers works as follows.
- Retreive the received message from the buffer
- Check that the message type is RTM_GETADDR because the only currently
  supported one is RTM_GETADDR
- Check that ifa_family of the message is either AF_UNSPEC or AF_INET
  because we support only IPv4 addresses
- Check that the rest of the fields are zeroes because they are
  unsupported
- Send the message containing all the interfaces

Each node in Shadow has only two interfaces: lo and eth0. The lo one has
the prefix of 8 and the eth0 one has the prefix of 24.

The detail of the interfaces is similar to the one in the getifaddrs
syscall handler.