Make PartialEq consistent with PartialOrd and Ord impl
This was an easy fix because we are using the priority u64 values
as unique ids in the priority queue, so we don't have to worry
about different epoll entries using the same priority key.
In other words, comparing the priority values yiedls the same
result as comparing the keys.