Implement Netlink getsockname syscall
Previously we have only the is_bound field in InitialState which is used
to indicate whether the socket is already bound or not. Now we want to
implement getsockname which requires us to return the bound address, so
we need to change from the boolean is_bound to the concrete bound_addr
field.