Handle and keep strace and log files as std File objects
When implementing exec, we will create a new `ManagedThread`, and want
to have that append to the same strace and log files as its predecessor.
Keeping around a File object for the log file and passing it through by
descriptor should be relatively easy to maintain, and more robust than
keeping a path and trying to reopen the same file to append to.

We were previously using nix to create these files with `O_CLOEXEC`, but
it looks like `std` uses `O_CLOEXEC` as well; it just doesn't document
it. So we use `std`, with an added check to verify that it really did
use `O_CLOEXEC`.