syscall-logger: log blocked syscalls
Previously if a syscall blocked, it wasn't recorded in the strace file
until if and when it ran again and completed. This saves some work of
logging the syscall and parameters multiple times, but makes it harder
to see why control switched to another thread.

e.g.

```
00:00:01.000000000 [tid 1000] clone(17, 0x0, 0x0, 0x0, 0x0) = 1001
00:00:01.000000000 [tid 1001] write(4, "*", 1) = 1
00:00:01.000000000 [tid 1001] exit_group(...) = <native>
00:00:01.000000000 [tid 1000] read(3, 0x7fffffffe707, 1) = 1
```

->

```
00:00:01.000000000 [tid 1000] clone(17, 0x0, 0x0, 0x0, 0x0) = 1001
00:00:01.000000000 [tid 1000] read(3, 0x7fffffffe707, 1) = <blocked>
00:00:01.000000000 [tid 1001] write(4, "*", 1) = 1
00:00:01.000000000 [tid 1001] exit_group(...) = <native>
00:00:01.000000000 [tid 1000] read(3, 0x7fffffffe707, 1) = 1
```

It's also helpful in the (unmerged) implementation of execve, which
"blocks" and never returns.